Fix IocProvider for netcore
