+++ Modules/MinoriEditorShell/MinoriEditorShell.csproj
@@ -38,8 +38,6 @@ This uses AvalonDock and has an MVVM architecture based on MvvmCross.</Descripti
         <PackageReference Include="System.ComponentModel.Composition" Version="4.7.0" />
         <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
         <PackageReference Include="System.Drawing.Common" Version="4.7.0" />
-
-        <PackageReference Include="System.ComponentModel" Version="4.3.0" />
     </ItemGroup>
 
     <ItemGroup>
