+++ Tests/MinoriEditorShell.RibbonTests.Wpf/MinoriEditorShell.RibbonTests.Wpf.csproj
@@ -51,9 +51,6 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.CodeCoverage">
-      <Version>16.7.0</Version>
-    </PackageReference>
     <PackageReference Include="MSTest.TestAdapter">
       <Version>2.1.2</Version>
     </PackageReference>
